Rename ScalarHasValue to HasValue #213
